requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
pymongo==4.6.0
python-dotenv==1.0.0
Werkzeug==2.3.7